nav footer 변경
