+++ src/components/Footer/Footer.jsx
@@ -1,0 +1,16 @@
+import {FootDiv,FootItem} from './FooterStyle'
+
+function Footer() {
+    return (
+        <FootDiv>
+            <FootItem>
+                노션링크
+            </FootItem>
+            <FootItem>
+                깃링크
+            </FootItem>
+        </FootDiv>
+    )
+}
+
+export default Footer;
